Add files via upload
These are my personal notes which is very targeted towards passing CC examination from ISC2. There are two Tabs below in the excel file.  Please understand the table and all the description upside down and there will be no chance of failing. The points mentioned are enough for passing this exam. On more trick, I would like to narrate:
1. Understand the table and the points very well.
2. Translate the understanding in the form of Question and Answers. The best way to generate Questions and Answers is by highlighting the points you understood and using AI- Chat GPT or Bing or Bard asking, "to generate 10 0r 15 or 20 (of your choice) multiple type questions in CC ISC2 style". Asking for answers as well. But please be careful about answers as they may not be accurate, but notes are there to validate.
3. Practice it again and again till you achieve 80% and after that you are ready. :)
All the best. Roy